# openapi.yaml
openapi: 3.1.0
info:
  title: LeanStock Inventory API
  description: LeanStock backend with JWT auth/RBAC, email verification, password
    reset, multi-tenant catalog, atomic inventory transfers, Redis reservations, Gmail
    OAuth2 email delivery, and scheduled dead-stock decay.
  version: 1.0.0
paths:
...
  /v1/warehouses:
    get:
      tags:
      - Warehouses
      summary: Read Warehouses
      operationId: read_warehouses_v1_warehouses_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/WarehouseResponse'
                type: array
                title: Response Read Warehouses V1 Warehouses Get
      security:
      - HTTPBearer: []
    post:
      tags:
      - Warehouses
      summary: Create Warehouse Endpoint
      operationId: create_warehouse_endpoint_v1_warehouses_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WarehouseCreateRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WarehouseResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    WarehouseCreateRequest:
      properties:
        name:
          type: string
          maxLength: 255
          minLength: 2
          title: Name
        location:
          anyOf:
          - type: string
            maxLength: 500
          - type: 'null'
          title: Location
      type: object
      required:
      - name
      title: WarehouseCreateRequest
    WarehouseResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        tenant_id:
          type: string
          format: uuid
          title: Tenant Id
        name:
          type: string
          title: Name
        location:
          anyOf:
          - type: string
          - type: 'null'
          title: Location
        is_active:
          type: boolean
          title: Is Active
      type: object
      required:
      - id
      - tenant_id
      - name
      - location
      - is_active
      title: WarehouseResponse
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer